requirements.txt
networkx==2.8
numpy==1.19.5
Pillow==9.1.0
scikit_learn==1.0.2
scipy==1.8.0
tensorflow==2.6.0
